+++ packages/component/src/Avatar/ImageAvatar.tsx
@@ -1,11 +1,10 @@
 import { hooks } from 'botframework-webchat-api';
 import classNames from 'classnames';
-import PropTypes from 'prop-types';
-import React from 'react';
+import React, { memo } from 'react';
 
-import CroppedImage from '../Utils/CroppedImage';
-import useStyleSet from '../hooks/useStyleSet';
+import FixedWidthImage from '../Utils/FixedWidthImage';
 import useStyleToEmotionObject from '../hooks/internal/useStyleToEmotionObject';
+import useStyleSet from '../hooks/useStyleSet';
 
 const { useAvatarForBot, useAvatarForUser } = hooks;
 
@@ -15,7 +14,7 @@ const ROOT_STYLE = {
   }
 };
 
-const ImageAvatar = ({ fromUser }) => {
+const ImageAvatar = memo(({ fromUser }: Readonly<{ fromUser: boolean }>) => {
   const [{ image: avatarImageForBot }] = useAvatarForBot();
   const [{ image: avatarImageForUser }] = useAvatarForUser();
   const [{ imageAvatar: imageAvatarStyleSet }] = useStyleSet();
@@ -26,24 +25,14 @@ const ImageAvatar = ({ fromUser }) => {
   return (
     !!avatarImage && (
       <div className={classNames('webchat__imageAvatar', rootClassName, imageAvatarStyleSet + '')}>
-        <CroppedImage
+        <FixedWidthImage
           alt=""
           className="webchat__imageAvatar__image"
-          height="100%"
           src={fromUser ? avatarImageForUser : avatarImageForBot}
-          width="100%"
         />
       </div>
     )
   );
-};
-
-ImageAvatar.defaultProps = {
-  fromUser: false
-};
-
-ImageAvatar.propTypes = {
-  fromUser: PropTypes.bool
-};
+});
 
 export default ImageAvatar;